fmp4: use sync for serializing boxes
Segments tend to be fairly small except for `mdat`, which is as
efficient to copy as a `Vec`, and we will always be serializing into
`Vec`s, so this won't block the async runtime too much. It's also more
ergonomic to make a custom `Stream` of segments when you don't have to
call async functions, since `Stream::poll_next` is sync.